requirements.txt
# Core NLP & AI
transformers==4.41.1
sentence-transformers==2.7.0
torch>=2.0.0
#faiss-cpu==1.7.4

# Whisper for speech-to-text
openai-whisper==20231117

# OCR for images
pytesseract==0.3.10
Pillow==10.3.0

# Audio/Video processing
moviepy==1.0.3

# Streamlit UI
streamlit==1.35.0

# Backend (for future admin dashboard, evidence handling)
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.3

# PDF/Text tools (optional if needed)
pdfplumber==0.10.4

# Others
numpy>=1.24.0
scikit-learn>=1.4.0
tqdm>=4.66.0
python-multipart==0.0.9
